Heart Disease Prediction
Heart disease prediction is a project that aims to use data and machine learning algorithms to predict the likelihood of an individual developing heart disease. This can be done by analyzing various risk factors such as age, blood pressure, cholesterol levels, and medical history, as well as lifestyle factors such as diet, exercise, and smoking habits. The ultimate goal of the project is to identify individuals at high risk of heart disease and intervene early to prevent the onset of the disease. This can be done by using various machine learning algorithms such as logistic regression, decision tree, Random Forest, and Naive Bayes.